Create Evaluating a Manufacturing Process_ SQL Code
This was a project completed to support a team wanting to improve how they monitor and control a manufacturing process. 

By implementing a statistical process control, I was able to create an alert that flags whether the height of a product is within the control limits for each operator, thereby helping managers ensure that product heights were between the highest acceptable height for the parts, and lowest acceptable height for the parts. 

Using SQL window functions and nested queries, I analysed historical manufacturing data to define this acceptable range and identified any points in the process that fell outside of the range and therefore require adjustments. This will ensure a smooth running manufacturing process consistently making high-quality products.